Added reactions_and_edits.py and reaction_roles.py
So basically the reactions_and_edits file makes the bot listen to edits made on message sent to the server and shows that so and so edited their message, showing the message before and after.

The reactions_roles file tells the bot to listen to certain or selected reaction on messages by users and assigns or rewards a role to a person based on their reaction on particular message.